First stable release :tada:
This package has been in beta for a while since I never really got the time to fully wrap it up. On the other hand we're in a place where we haven't made any significant changes for a while, and Umbraco 7 is now EOL. So as a result of this, I'm now releasing the first (and probably last) stable release for Umbraco 7.